move functions to utils.js
